test: Update tests
Add test for listing versions to apply

It's a helper function, though I'm having trouble with the algorithm.
It is based off the "comm" command from GNU coreutils, which compares
sorted lines and selects certain ones. This function will be a central
part in choosing which migration to run, so it's very important that it
works properly.