refactor: Remove 2 fields from MigrationsConf
We cannot interpolate identifier fields into SQL queries (at least in
Postgres). An alternative would be to use golang's fmt.Sprintf function
to do the interpolation and then pass to the sql library. But doing that
is a security compromise because it foregoes the benefits of automatic
escaping. Instead, I've chosen to hardcode the identifiers:
"schema_migrations", "migration_id". This is a small feature to retract
(customize the name of db migration table and column), but I don't think
most people would care that you can't customize it anymore and a small
improvement in security is good.