refactor: Change Driver constructor, rename local vars and a function
More flexibility in testing now that we can pass in connection params.

refactor: change, normalize local variable names

The old variable "path" meant a directory that a file resided in. Not
the path to the file. Use same convention throughout.

refactor: rename ListAllAVailableMutations

refactor: export ListAppliedVersions function

It's part of the Driver interface, so this seems consistent with that,
but prepending "List" in the name to match similar functions.