feat: Update runMutation function and Driver interface
Now this function will use the Driver to create the schema migrations
table if needed and update it after running the mutation. Changed the
name of ApplyMigration function in Driver interface b/c the runMutation
method is actually "applying it", whereas UpdateSchemaMigrations does
just that. We're assuming that the sql in the executed file is written
to work for the given database driver.