refactor: Rename types, methods, variables; update function signatures
Rename pathToMigration variable. More vague is actually more helpful in
this case b/c the name "migration" doesn't match this package's
semantics.

refactor: rename MigrateOne to ApplyMigration

Closer to the dogfish's API, I guess.

refactor: Change field name & parameter for Migration type

More distinguished from "direction".

refactor: change NewPostgres signature

functionally is the same, but organizes better documentation.

refactor: Rename Migration to MigrationParams

Looking at the code in a little more macro sense than before. The name
Migration is kind of confusing because this type is a wrapper for the
thing that does the work. Also, it's most useful when generating the
files. Otherwise, it should be out of the way. Update godoc comments.

refactor: rename Mutation->Migration; Modification->Mutation

I think "Migration" will probably resonate more with people, since this
is a migration manager. Using Migration for the interface because it's
higher-level. Scrapping Modification because it sounds square and
Mutation sounds cool, so keeping it around; except now it's the name of
the concrete implementation instead of abstract interface. Update
variables and godoc as needed.

docs: update godoc stuff