refactor: condense exported API
It's too bloated and it should only expose identifiers used by the
bin/main.go file. Meanwhile, we still need to be able to test some of
the (now unexported) helpers. So those tests should be in a file other
than godfish_test.

Mutation and Postgres aren't part of any exported function signatures,
so they don't need to be exposed. As uncle bob says, hide the concrete
implementations and use the abstract interfaces. Changed name of
PGParams type to PostgresPaprams to make it more obvious. This needs to
be exported so you can use the NewDriver function.