test: split up test files, update tests
Need to be able to test some unexported helper functions. The current
test file, godfish_test is its own package (blackbox testing).

Add thorough testing of Driver interface. The tests touch on basic
functionality, except for error handling. Added several helper functions
to deal with I/O minutia. Not sure if other Driver implementations will
handle it too well, so I kept the SQL simple. There is now a one second
lag between generating each migration file. This is annoying, but it's
so each timestamp is different from the previous. When choosing a
migration file to run, it uses the timestamp only, not the migration's
name. The same timestamp scenario is unlikely to happen in real life,
but it happens all the time in tests. If the test suite gets annoyingly
slow, it would be a good time to rethink file selection, but it's not
the biggest priority at the moment.

test: update tests after type renames