refactor: mv source files to /godfish and bin to /
Should be able to do "go install" to generate the CLI binary, while
letting the command name to be "godfish", not "bin". Moved the source
files to the godfish/ directory, the same name as the binary after go
install. I don't think this will be a problem though because the go
toolset treats package main differently.