refactor: change filename structure
While writing documentation I was noting some difference compared to
dogfish and I was trying to think of a reason to justify the pattern for
generated filenames. In dogfish it's "direction-version-name" and here
it's "version.direction.name". Note, the delimiter is different as well.
I couldn't think of a good reason, it was a rather arbitrary decision to
begin with. So, I'm making another arbitrary decision to have the
generated filenames match the same structure as in dogfish. However,
keeping the terms "forward", "reverse" as they are since they have
meaning for godfish.

test: Update tests after filename structure change

docs: Update README