feat: Update MigrationsConf type, add Init function and conf opts
Add driver name field to MigrationsConf type, so we don't have to pass
it in for every CLI invocation.
Adding the Init function brings in two more stdlib dependencies. This
function could easily go into the CLI, but I'm putting it in the source
package for better testing.
Removed MigrationsConf field from postgres b/c it was never used. At
the start of the project, it was to customize the table and column name
for schema migrations, but that's not worth the effort.

feat(bin): Add conf, files options, update init command

Also set default values for connecting with a driver

test: Add test for Init