refactor: Update NewDriver function, DSNParams interface
Detach the main constructor function from any concrete implementations.
This extra layer of indirection will enable more flexibility to add more
Drivers in the future and, with the help of build tags and code generation,
help make builds smaller. Preemptively have added a MigrationsConf
argument to the constructors. This is kind of a grab bag of options you
might put in a conf file and might want to pass on to the Driver. Making
that a pointer, rather than value makes it easier to "opt out". Removed
the redundant string driverName argument.