refactor: Move driver implementations into own packages
There are some design goals in this project that can't be met without
making this split. They are: interface with many RDBMSs and to depend
on as little as possible (outside of the standard library). If I'm
working on a FooDB project, I shouldn't need to also set up BarDB just
to run the test suite or to even use the tool. Save that for continuous
integration environments.

refactor: rename types

test: Set up driver tests using internal package

chore: lint godoc in godfish package