test: fix Driver tests
Most of the tests were operating on no migration files at all. This
would catch very grave errors, but is not a realistic scenario. Add
setup and teardown steps for every subtest so there is some data.

We can indirectly test that the Driver's AppliedVersions method returns
the correct error value by removing the unit tests that create the
schema migration table. It should be called as needed by running a
migration. If it's not, then listing applied versions will fail.

Add godoc comment.