refactor: Unexport some errors (breaking change)
Take advantage of go 1.13 error wrapping, chains.

Only need to export things used by a driver implementation. The driver
tests did look at this, but the results have nothing to do with how a
driver is implemented. This is a breaking change. However, I don't
think anybody else uses this package, so I'd welcome the damage, if any.