fix: mysql migration bugs & simplify run migration
Was seriously considering having the godfish.runMigration function run
multiple-statement migrations in a transaction. However, it looks like
doing this from godfish is kind of complicated and would require
interface changes, particularly with changes needed to continue testing
with internal/stub.Driver. It looks like this problem only appears in
the mysql driver, so solve it there and don't push a solution into the
core library.

fix(mysql): multiple-statement migrations

Attempt to implement rollbacks in case of an error, although I'm not
sure there is a way to actually "undo" successful DDL statements with
mysql when a subsequent statement in the same migration file errors
out. By default, DDL statements commit implicitly in mysql.
https://dev.mysql.com/doc/refman/5.7/en/implicit-commit.html

This means that some statements in the migration would be applied, some
not (the first one that had an error). Though, the "version" won't be
recorded to the schema migrations table. This is not a great solution,
but I don't know how to make it better.

Also fix migration file parsing issues.

test: Update tests