fix: Connect using DB_DSN env var
The DSN formation scheme is too rigid. There are so many possible
connection parameters, that supporting them all is just infeasible. It's
better to use the same piece of data for connecting to the DB in the
first place, that is the data source name. The standard library already
works this way and it should be a little easier to use godfish now that
there are fewer environment variables to manage.