build: Cleanup Makefile
Draws a boundary between each Driver implementation in the test suite.
The goal is to have less assumptions about the host: it should be able
to build the code, but doesn't need a working postgres and/or mysql
installation to get most of the way there. Testing the code against a
live DB is recommended, but it's separated now.