refactor: Make output for cmd help, adhoc data more consistent
Usage funcs in pkg cmd would write some stuff to stdout, some to stderr.
Make this consistent. Going with stdout for usage funcs to make it
easier to capture help in a file.

Specifying outputs in the library code helps make the Info subcommand
cleaner. If you've selected the json format output, you're piping to jq
and there's an error, then invalid json is written to stdout. So, write
that adhoc stuff to stderr.