Apply security-related changes, update codecov config
Am trying out github.com/securego/gosec. Ran v2.10.0 on this project and
fixed the issues.

Change codecov config to disable status checks. I would prefer not to
have the red X next to the commit hash when the code coverage takes a
dip. I evaluate code coverage on a case-by-case basis. In this case,
code changes are mostly for closing files, which can be difficult to
test anyways. It's okay if coverage dips every now and then as long as
it's not consistently going downwards.