build: Add goreleaser
Automates binary builds for multiple platforms. More info on this tool
is at https://goreleaser.com.

The config file may also set env vars, but it doesn't meet my needs for
setting up LDFLAGS, so I'm keeping that in the Makefile. For
consistency's sake, the other env vars are also in the Makefile.