feat: Add colorized logging format
Also make this the default output format.

About library choice, I also looked at gitlab.com/greyxor/slogor, but
found that its coloring was too bright too often. IMO, it's better when
the field names have a color, and the field values are the regular text
color. In slogor, this was inverted, and it was too much. I chose
github.com/lmittmann/tint because it does not overdo the colors for the
kind of values I'd expect to see from this tool (ie: path to a file),
and because IMO its API is simpler.