build statically-linked binaries
Do this so that there is less dependency on the host OS to have
particular libraries available, such as glibc.

When building release binaries, also remove debugging symbols. This is
the go tool link flag, -s.
But keep them if building binaries when using the Makefile directly, as
that is more likely to be a development environment.